requirements.txt
apturl==0.5.2
asgiref==3.8.1
async-timeout==4.0.3
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-cors-headers==4.4.0
django-ipware==7.0.1
djangorestframework==3.15.2
duplicity==0.8.21
evdev==1.4.0
fasteners==0.14.1
filelock==3.15.4
future==0.18.2
hiredis==3.0.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
kavenegar==1.1.2
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
mutagen==1.45.1
mysqlclient==2.1.1
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
panadas==0.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.2.2
ply==3.11
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
PuLP==2.8.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
Pyomo==6.7.1
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-ipware==3.0.0
python-magic==0.4.24
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
redis==5.0.8
reportlab==3.6.8
requests==2.25.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
setproctitle==1.2.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.5.0
ssh-import-id==5.11
systemd-python==234
termcolor==2.4.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.26.3
wadllib==1.3.6
webencodings==0.5.1
websockets==9.1
WoeUSB-ng==0.2.12
wxPython==4.0.7
xdg==5
xkit==0.0.0
yt-dlp==2022.4.8
zipp==1.0.0
